requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --strip-extras
#
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
idna==3.4
    # via requests
mypy==1.7.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via mypy
packaging==23.2
    # via build
pip-tools==7.3.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
ruff==0.3.2
    # via -r requirements.in
types-requests==2.31.0.10
    # via -r requirements.in
typing-extensions==4.8.0
    # via mypy
urllib3==2.1.0
    # via
    #   requests
    #   types-requests
wheel==0.41.3
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
